# doc/api.yml
openapi: 3.0.3
info:
  version: 'latest'
  title: Bookwel-api
servers:
  - url: 'http://13.39.144.179:3000'
security:
  - BearerAuth: [ ]
paths:
...
  /users:
    get:
      tags:
        - Users
      summary: get all users
      operationId: getAllUsers
      responses:
        200:
          description: the users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
        400:
          $ref: "#/components/responses/400"
        500:
          $ref: "#/components/responses/500"
        404:
          $ref: "#/components/responses/404"
    post:
      tags:
        - Users
      summary: CrupdateUser
      operationId: crupdateUser
      parameters:
        - name: userId
          in: path
          schema:
            type: string
          required: true
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateUser"
      responses:
        200:
          description: the created user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        400:
          $ref: "#/components/responses/400"
        500:
          $ref: "#/components/responses/500"
        404:
          $ref: "#/components/responses/404"
  /users/{userId}:
    get:
      tags:
        - Users
      summary: getUserById
      operationId: getUserById
      parameters:
        - name: userId
          in: path
          schema:
            type: string
          required: true
      responses:
        200:
          description: the user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        400:
          $ref: "#/components/responses/400"
        500:
          $ref: "#/components/responses/500"
        404:
          $ref: "#/components/responses/404"
  /users/{userId}/profile:
    put:
      tags:
        - Users
      summary: update user profile
      operationId: updateUserProfile
      parameters:
        - name: userId
          in: path
          schema:
            type: string
          required: true
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserProfile"
      responses:
        200:
          description: the created user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        400:
          $ref: "#/components/responses/400"
        500:
          $ref: "#/components/responses/500"
        404:
          $ref: "#/components/responses/404"
components:
  responses:
    400:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/BadRequestException'
...
    404:
      description: Not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ResourceNotFoundException'
...
    '500':
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/InternalServerException'
  schemas:
    Exception:
      type: object
      properties:
        type:
          type: string
        message:
          type: string
    InternalServerException:
      allOf:
        - $ref: '#/components/schemas/Exception'
      example:
        type: InternalServerException
        message: Unexpected error
    BadRequestException:
      allOf:
        - $ref: '#/components/schemas/Exception'
      example:
        type: BadRequestException
        message: Bad request
...
    ResourceNotFoundException:
      allOf:
        - $ref: '#/components/schemas/Exception'
      example:
        type: ResourceNotFoundException
        message: Resource of type <T> identified by <I> not found
...
    UserProfile:
      type: object
      properties:
        first_name:
          type: string
        last_name:
          type: string
        email:
          type: string
          format: email
    User:
      type: object
      allOf:
        - $ref: "#/components/schemas/CreateUser"
      properties:
        profile:
          $ref: "#/components/schemas/UserProfile"
        status:
          $ref: "#/components/schemas/UserStatus"
    UserStatus:
      type: string
      enum:
        - ADMIN
        - CLIENT
...
    CreateUser:
      type: object
      properties:
        id:
          type: string
        firebaseId:
          type: string
...