# src/main/java/api/prog5/bookwel/endpoint/rest/exception/BadRequestException.java
package api.prog5.bookwel.endpoint.rest.exception;

public class BadRequestException extends ApiException {
  public BadRequestException(String message) {
    super(ExceptionType.CLIENT_EXCEPTION, message);
  }
}